Move scene graph node popout menu into a spearate component
