+++ src/panels/Scene/SceneGraphNode/SceneGraphNodeHeader.tsx
@@ -1,17 +1,15 @@
-import { ActionIcon, Button, Divider, Group, Menu, Text } from '@mantine/core';
+import { Button, Group, Text } from '@mantine/core';
 
 import { useGetPropertyOwner } from '@/api/hooks';
 import { NodeNavigationButton } from '@/components/NodeNavigationButton/NodeNavigationButton';
 import { PropertyOwnerVisibilityCheckbox } from '@/components/PropertyOwner/VisiblityCheckbox';
 import { ThreePartHeader } from '@/components/ThreePartHeader/ThreePartHeader';
-import { OpenInNewIcon, VerticalDotsIcon } from '@/icons/icons';
 import { useAppSelector } from '@/redux/hooks';
-import { IconSize, NavigationType } from '@/types/enums';
+import { NavigationType } from '@/types/enums';
 import { Uri } from '@/types/types';
 import { displayName, sgnRenderableUri } from '@/util/propertyTreeHelpers';
-import { useWindowLayoutProvider } from '@/windowmanagement/WindowLayout/hooks';
 
-import { SceneGraphNodeView } from './SceneGraphNodeView';
+import { SceneGraphNodeMoreMenu } from './SceneGraphNodeMoreMenu';
 
 interface Props {
   uri: Uri;
@@ -27,21 +25,11 @@ export function SceneGraphNodeHeader({ uri, label, onClick }: Props) {
     return state.propertyOwners.propertyOwners[renderableUri] !== undefined;
   });
 
-  const { addWindow } = useWindowLayoutProvider();
-
   if (!propertyOwner) {
     return <></>;
   }
 
   const name = label ?? displayName(propertyOwner);
-
-  function openInNewWindow() {
-    addWindow(<SceneGraphNodeView uri={uri} />, {
-      id: 'sgn-' + uri,
-      title: name,
-      position: 'float'
-    });
-  }
 
   return (
     <ThreePartHeader
@@ -77,42 +65,7 @@ export function SceneGraphNodeHeader({ uri, label, onClick }: Props) {
             type={NavigationType.focus}
             identifier={propertyOwner.identifier}
           />
-          <Menu position={'right-start'}>
-            <Menu.Target>
-              <ActionIcon size={'sm'} variant={'light'}>
-                <VerticalDotsIcon />
-              </ActionIcon>
-            </Menu.Target>
-            <Menu.Dropdown>
-              <Menu.Label>{name}</Menu.Label>
-              <Button
-                onClick={openInNewWindow}
-                leftSection={<OpenInNewIcon size={IconSize.sm} />}
-              >
-                Pop out
-              </Button>
-              <Divider m={'xs'} />
-              <Group gap={'xs'}>
-                <NodeNavigationButton
-                  type={NavigationType.fly}
-                  identifier={propertyOwner.identifier}
-                  showLabel
-                />
-                <NodeNavigationButton
-                  type={NavigationType.jump}
-                  identifier={propertyOwner.identifier}
-                  showLabel
-                />
-              </Group>
-              <Group gap={'xs'}>
-                <NodeNavigationButton
-                  type={NavigationType.frame}
-                  identifier={propertyOwner.identifier}
-                  showLabel
-                />
-              </Group>
-            </Menu.Dropdown>
-          </Menu>
+          <SceneGraphNodeMoreMenu uri={uri} />
         </Group>
       }
     />
